// src/main/model/rawPocketArticle.ts
// tslint:disable:interface-name
// tslint:disable:variable-name
export default interface RawPocketArticle {
    item_id?: string;
    resolved_id?: string;
    given_url?: string;
    given_title?: string;
    favorite?: string;
    status?: string;
    time_added?: string;
    time_updated?: string;
    time_read?: string;
    time_favorited?: string;
    sort_id?: string;
    resolved_title?: string;
    resolved_url?: string;
    excerpt?: string;
    is_article?: string;
    is_index?: string;
    has_video?: string;
    has_image?: string;
    word_count?: string;
}